tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0